Migrate mappers to extension functions for better readability and less boilerplate, remove mapper APIs in core-infra.
