+++ data/src/main/java/ychescale9/releaseprobe/data/artifact/fetcher/ArtifactGroupsWithArtifactsFetcher.kt
@@ -4,13 +4,12 @@ import com.nytimes.android.external.store3.base.Fetcher
 import com.nytimes.android.external.store3.base.impl.BarCode
 import io.reactivex.Single
 import ychescale9.infra.SchedulerProvider
-import ychescale9.releaseprobe.data.artifact.mapper.ArtifactGroupWithArtifactsDtosToEntity
+import ychescale9.releaseprobe.data.artifact.mapper.toEntity
 import ychescale9.releaseprobe.persistence.artifact.entity.ArtifactGroupWithArtifactsEntity
 import ychescale9.releaseprobe.remote.artifact.api.GoogleMavenService
 
 class ArtifactGroupsWithArtifactsFetcher(
     private val googleMavenService: GoogleMavenService,
-    private val mapper: ArtifactGroupWithArtifactsDtosToEntity,
     private val schedulerProvider: SchedulerProvider
 ) : Fetcher<List<ArtifactGroupWithArtifactsEntity>, BarCode> {
 
@@ -23,7 +22,7 @@ class ArtifactGroupsWithArtifactsFetcher(
                     googleMavenService.fetchArtifactsByGroupId(path)
                             .subscribeOn(schedulerProvider.io())
                             .map { artifacts ->
-                                mapper.transform(Pair(artifactGroup, artifacts))
+                                Pair(artifactGroup, artifacts).toEntity()
                             }
                             .toObservable()
                 }
