+++ data/src/main/java/ychescale9/releaseprobe/data/artifactcollection/mapper/ArtifactCollectionMappers.kt
@@ -1,0 +1,12 @@
+package ychescale9.releaseprobe.data.artifactcollection.mapper
+
+import ychescale9.releaseprobe.domain.artifactcollection.model.ArtifactCollection
+import ychescale9.releaseprobe.persistence.artifactcollection.entity.ArtifactCollectionEntity
+
+fun ArtifactCollectionEntity.toModel(): ArtifactCollection =
+    ArtifactCollection(
+        name = name,
+        description = description,
+        themeColor = themeColor,
+        keywords = keywords
+    )
