+++ detekt.yml
@@ -20,7 +20,7 @@ test-pattern: # Configure exclusions for test sources
     - 'ForEachOnRange'
 
 build:
-  maxIssues: 5
+  maxIssues: 0
   weights:
     # complexity: 2
     # LongParameterList: 1
