--- library/core-infra/src/main/java/ychescale9/infra/extension/CollectionExtensions.kt
@@ -1,8 +1,0 @@
-package ychescale9.infra.extension
-
-/**
- * Compare whether 2 collections are the same ignoring order
- */
-infix fun <T> Collection<T>.sameContentWith(collection: Collection<T>) = collection.let {
-    this.size == it.size && this.containsAll(it)
-}
