--- library/core-infra/src/main/java/ychescale9/infra/mapper/BaseItemMapper.kt
@@ -1,5 +1,0 @@
-package ychescale9.infra.mapper
-
-interface BaseItemMapper<From, To> {
-    fun transform(item: From, vararg params: Any): To
-}
