--- library/core-infra/src/main/java/ychescale9/infra/mapper/DataMapper.kt
@@ -1,6 +1,0 @@
-package ychescale9.infra.mapper
-
-interface DataMapper<From, To> {
-    fun transform(model: From, vararg params: Any): To
-    fun transform(models: Collection<From>): Collection<To>
-}
